Refactor report handling in ReportDetailPage and ActionButtons components for improved clarity and functionality
